+++ app/api/v1/endpoints/adventures.py
@@ -6,7 +6,7 @@ from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 
 from app.schemas.adventure import AdventureUpdateSchema, AdventureCreateSchema, AdventureSchema
-from app.api.dependencies import get_db
+from app.api.v1.dependencies import get_db
 from app.crud import adventure as crud
