+++ app/api/v1/endpoints/error_details.py
