+++ app/api/v1/endpoints/users.py
@@ -6,9 +6,9 @@ from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 
 from app.schemas.user import UserSchema, UserCreateSchema, UserUpdateSchema
-from app.api.dependencies import get_db, get_current_user
+from app.api.v1.dependencies import get_db, get_current_user
 from app.crud import user as crud
-from app.api.endpoints import user_errors as error_details
+from app.api.v1.endpoints import user_errors as error_details
 from app.models.user import User
 
 
@@ -16,7 +16,7 @@ router = APIRouter()
 
 
 @router.get('/', response_model=list[UserSchema])
-def read_all(db: Session = Depends(get_db)):
+def read_all(db: Session = Depends(get_db), current_user: User = Depends(get_current_user)):
     """Read users"""
     return crud.get_users(db)
 
@@ -42,8 +42,12 @@ def get_current_user_info(current_user=Depends(get_current_user)):
 
 
 @router.get('/{user_id}', response_model=UserSchema)
-def read_one(user_id: int, db: Session = Depends(get_db)):
+def read_one(user_id: int,
+             db: Session = Depends(get_db),
+             current_user: User = Depends(get_current_user)):
     """Get user by id"""
+    if not current_user.user_id == user_id:
+        raise HTTPException(status_code=403, detail=error_details.USER_NOT_AUTHORIZED)
     user = crud.get_user_by_id(user_id=user_id, db=db)
     if user is None:
         raise HTTPException(status_code=404)
